Add stub ShaderGraphStencilMock, split mock classes into own files
