+++ com.unity.sg2/Tests/GraphUI/ShaderGraphStencilMock.cs
@@ -1,0 +1,7 @@
+namespace UnityEditor.ShaderGraph.GraphUI.UnitTests
+{
+    class ShaderGraphStencilMock : ShaderGraphStencil
+    {
+        // Stencil overrides for testing can go here.
+    }
+}
