+++ com.unity.sg2/Tests/GraphUI/ShaderGraphStencilMock.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1f42d253026f4799b695e4c22cf84b38
+timeCreated: 1674851789
